EditSuccess responder now follows new conventions.
Namely, extending the PostDao instance it returns with the
PostDao::Presentation module, which contains methods such as `#draft?`.

This simplifies the bare PostDao class *significantly*.

RSpec: 885 examples, 0 failures, 1 pending
RuboCop: 279 files inspected, no offenses detected
